# backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.ClientCompanyListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.ClientCompanyResponse'
        type: array
      meta:
        properties:
          current_page:
            type: integer
          last_page:
            type: integer
          total:
            type: integer
        type: object
    type: object
  handlers.ClientCompanyResponse:
    properties:
      address:
        type: string
      archived_at:
        type: string
      city:
        type: string
      clients:
        items:
          $ref: '#/definitions/handlers.ClientUserResponse'
        type: array
      country_id:
        type: integer
      currency_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      web:
        type: string
    type: object
  handlers.ClientUserResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  handlers.CountryResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Indonesia
        type: string
    type: object
  handlers.CurrencyResponse:
    properties:
      code:
        example: IDR
        type: string
      decimals:
        example: 2
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Rupiah
        type: string
      symbol:
        example: Rp
        type: string
    type: object
  handlers.GoogleLoginRequest:
    properties:
      credential:
        type: string
    required:
    - credential
    type: object
  handlers.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: password
        type: string
    required:
    - email
    - password
    type: object
  handlers.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        example: 900
        type: integer
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/handlers.UserResponse'
    type: object
  handlers.OwnerCompanyResponse:
    properties:
      address:
        type: string
      city:
        type: string
      country_id:
        type: integer
      currency_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      web:
        type: string
    type: object
  handlers.PermissionResponse:
    properties:
      category:
        example: Roles & Permissions
        type: string
      description:
        example: Change which permissions are assigned to a role
        type: string
      id:
        example: 1
        type: integer
      name:
        example: roles.manage
        type: string
    type: object
  handlers.RoleResponse:
    properties:
      id:
        example: 2
        type: integer
      name:
        example: manager
        type: string
      permissions:
        items:
          type: string
        type: array
      version:
        example: 1
        type: integer
    type: object
  handlers.UpdateRolePermissionsRequest:
    properties:
      permissions:
        items:
          type: string
        type: array
      version:
        type: integer
    required:
    - version
    type: object
  handlers.UserResponse:
    properties:
      email:
        example: admin@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Admin User
        type: string
      permissions:
        items:
          type: string
        type: array
      role:
        example: admin
        type: string
    type: object
info:
  contact: {}
  description: Backend API for the Construction Project Manager application.
  title: CPM API
  version: "0.1"
paths:
  /api/v1/auth/google:
    post:
      consumes:
      - application/json
      description: Verifies a Google Identity Services ID token and logs in an existing
        user (matched by google_id or email). Does not create new accounts.
      parameters:
      - description: Google ID token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in with Google
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns an access token; sets an httpOnly
        refresh-token cookie.
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Revokes the current refresh token and clears its cookie.
      responses:
        "204":
          description: No Content
      summary: Log out
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      description: Rotates the refresh-token cookie and returns a new access token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/client-companies:
    get:
      description: Paginated, searchable. Requires client_company.view.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Search by name or email
        in: query
        name: search
        type: string
      - description: Show archived instead of active
        in: query
        name: archived
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ClientCompanyListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List client companies
      tags:
      - client-companies
    post:
      consumes:
      - multipart/form-data
      description: multipart/form-data body (logo file is optional). Requires client_company.create.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ClientCompanyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a client company
      tags:
      - client-companies
  /api/v1/client-companies/{id}:
    delete:
      parameters:
      - description: Client company ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Archive a client company
      tags:
      - client-companies
    get:
      parameters:
      - description: Client company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ClientCompanyResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a client company
      tags:
      - client-companies
    put:
      consumes:
      - multipart/form-data
      description: multipart/form-data body (logo file is optional). Requires client_company.edit.
      parameters:
      - description: Client company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ClientCompanyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a client company
      tags:
      - client-companies
  /api/v1/client-companies/{id}/force:
    delete:
      parameters:
      - description: Client company ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Permanently delete a client company
      tags:
      - client-companies
  /api/v1/client-companies/{id}/restore:
    post:
      parameters:
      - description: Client company ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Restore an archived client company
      tags:
      - client-companies
  /api/v1/lookups/client-users:
    get:
      description: For the "assign clients" picker on the ClientCompany form.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ClientUserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List users with the client role
      tags:
      - client-companies
  /api/v1/lookups/countries:
    get:
      description: Static reference data, for populating dropdowns.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.CountryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List countries
      tags:
      - lookups
  /api/v1/lookups/currencies:
    get:
      description: Static reference data, for populating dropdowns.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.CurrencyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List currencies
      tags:
      - lookups
  /api/v1/me:
    get:
      description: Returns the authenticated user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
  /api/v1/owner-company:
    get:
      description: Returns the single owner company record. Requires owner_company.view.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.OwnerCompanyResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get owner company
      tags:
      - owner-company
    put:
      consumes:
      - multipart/form-data
      description: multipart/form-data body (logo file is optional). Requires owner_company.edit.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.OwnerCompanyResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update owner company
      tags:
      - owner-company
  /api/v1/permissions:
    get:
      description: Returns every permission in the catalog. Requires the roles.view
        permission.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.PermissionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all permissions
      tags:
      - permissions
  /api/v1/roles:
    get:
      description: Returns every role except admin (whose access can't be restricted),
        each with its assigned permission names. Requires the roles.view permission.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.RoleResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List manageable roles with their permissions
      tags:
      - permissions
  /api/v1/roles/{id}/permissions:
    put:
      consumes:
      - application/json
      description: Replaces the full set of permissions assigned to a role (admin's
        role can't be changed here). Requires the roles.manage permission.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission names to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RoleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a role's assigned permissions
      tags:
      - permissions
  /health:
    get:
      description: Returns service status, used for uptime checks and CI/CD deploy
        verification
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
      summary: Health check
      tags:
      - system
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"